ec_diffuser/requirements.txt
einops==0.7.0
gym==0.21.0
imageio==2.33.1
matplotlib==3.7.4
moviepy==1.0.3
numpy==1.21.0
opencv_python==4.9.0.80
Pillow==10.2.0
piqa==1.3.2
pytorch_lightning==2.1.3
pyvirtualdisplay==3.0
PyYAML==6.0.1
Requests==2.31.0
stable_baselines3==1.5.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
wandb==0.16.2
mujoco-py==2.1.2.14
scikit-video
typed-argument-parser